requirements.txt
python-telegram-bot==22.4
python-dotenv==1.0.0